Move const to utils
