+++ xawscf/pattern/function.py
@@ -1,6 +1,6 @@
 from ..utils import helper
 from ..utils.yaml import Custom
-from .const import SOURCES_BUCKET
+from ..utils.const import SOURCES_BUCKET
 from .utils import get_full_name, try_set_field, make_output
 from .base_resource import BaseResource
